Thin wrapper around CGAL to perform CSG operations on Polyhedra
This is an implementation of a single function csgop() which accepts two
Polyhedra and an operation identifier as arguments, performs the CSG
operation on the two polyhedra using CGAL and returns a data structure
representing the resulting Polyhedron.

The polyhedra data structures are converted into CGAL Polyhedron data
structure, using the incremental builder functionality. These polyhedron
are then converted to Nef_Polyhedron, which support CSG operations like
Union.

This code is meant to be compiled into a SharedLibrary, which can be
opened from python using ctypes.

[This code currently only does union, doing subtraction and
intersection is trivial and will be added soon.]