Python backend that opens native libcsgop.so using ctypes
This commit includes the `cadmium` module, which includes a backend that
interfaces with the native libcsgop.so using ctypes. It also defines
several datatypes that are used by python code to pass data back-n-forth
with the native backend. It also opens libc.so and defines a reference
to free(), which is used by python code to free the memory allocated by
the libcsgop.so.