Polyhedron class
Polyhedron class manages the vertices and faces of a polyhedron and
passes them back-n-forth through the backend to CGAL.

It overrides the __add__ method so that p1+p2 gives union of two
polyhedrons. It also has toSTL() method which exports it as an STL
binary data which can be saved to file and opened in STL viewers.