Simpler csgop implementation
This is an alternate simpler way of passing Polyhedron data to CGAL.
This approach doesn't use any complex data structures, but a plain
string to pass the OFF representation of the Polyhedron. The memory
management is entirely done in Python, no memory is allocated in the C
library, so no need to worry about memory leak, corruption problems. The
OFF string representation of polyhedron is passed as constant string
arguments to the C function csgop_simple.

The resultant Polyhedron of the CSG operation is not returned directly to
the caller, but is written to OFF file on filesystem. The name of the file
is agreed upon as <ID>.off, where ID is integer that is returned to the python
caller. Python caller then reads the polyhedron from the file (deleting it
subsequently).

This unusual method of returning polyhedron data is used to avoid
allocating memory in C library. The size of such memory is not known
before hand, so we cannot allocate it in python and then pass to C
library to write to it. Conventionally, passing data via filesystem is
much slower alternative to passing data in-memory. However in our case
the time spent in computation is large enough to ignore this tiny
performance difference.