Rename Polyhedron to Solid
The name Polyhedron may be misleading in the context of PythonOCC. This
base class is just meant to hold generic functions that are common
across all solids - affine transformations and export helper functions.
Therefore it's now renamed to Solid.

Solid is also exported for the cadmium users to derive classes from,
which will be helpful to create object oriented solid generation
scripts.