Get thickness of the text working
In order to get the desired thickness, we need to pre-calculate the
extents of the text, then calculate the scaling factor which we will
have to apply. Then we pre-multiply the desired thickness of the text.
This way when uniform scaling is applied at the end, the thickness gets
properly adjusted to desired thickness.

[This round-about way is needed because because non-uniform scaling
doesn't work quite well]