Implement BREP caching for Glyph solid creation
The cadmium library creates a cache directory at configurable location,
where it store the BREP files of Glyph solids it generates on the go.
These files are named after the signature of that glyph instance, which
include all the parameters that define the glyph.

Typically Glyph is used through the Text solid. In order to leverage the
caching behavior of Glyph, the Text object always creates Glyphs of
thickness 1 and subsequently scales them in Z as per the required
thickness.